update docs and internals organization
