+++ docs/source/functions.rst
