+++ docs/source/utils.rst
@@ -1,5 +1,5 @@
-utils module
-============
+Utility Functions
+=================
 
 .. automodule:: src.utils
    :members: fit_mlp_cv
